requirements.txt
#FOR PYTHON 3.10:
# Requirements.txt

# Core
opencv-python==4.9.0.80
numpy==1.26.4
av>=8.0.0
mediapipe==0.10.14
tqdm>=4.60.0

# Audio
scipy>=1.5.0
soundfile>=0.10.0

# PyTorch (local, compatible with pyannote.audio 2.1.1)
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0

# Pyannote (local)
pyannote.audio==2.1.1
pyannote.core==4.5.0
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3.0

# HuggingFace helpers (optional, for auth/files)
transformers==4.57.1
huggingface-hub<1.0

# Utilities
argparse

#NOTE: Only if you are using a .env file for hf token, install this library
python-dotenv


